Cleans up token and bankAccount modules
- Extracts normalizer into separate module
- Removes unnecessary routines
- Cleans up some syntax
