+++ lib/recurly/bank-account.js
@@ -3,12 +3,11 @@
  * Module dependencies.
  */
 
-var bind = require('bind');
 var each = require('each');
 var type = require('type');
 var index = require('indexof');
 var debug = require('debug')('recurly:bankAccount');
-var dom = require('../util/dom');
+var normalize = require('../util/normalize');
 var errors = require('../errors');
 
 /**
@@ -28,29 +27,29 @@ module.exports = {
  */
 
 var fields = [
-    'account_number',
-    'account_number_confirmation',
-    'routing_number',
-    'name_on_account',
-    'account_type',
-    'address1',
-    'address2',
-    'country',
-    'city',
-    'state',
-    'postal_code',
-    'phone',
-    'vat_number',
-    'token'
+  'account_number',
+  'account_number_confirmation',
+  'routing_number',
+  'name_on_account',
+  'account_type',
+  'address1',
+  'address2',
+  'country',
+  'city',
+  'state',
+  'postal_code',
+  'phone',
+  'vat_number',
+  'token'
 ];
 
 var requiredFields = [
-    'account_number',
-    'account_number_confirmation',
-    'routing_number',
-    'account_type',
-    'name_on_account',
-    'country'
+  'account_number',
+  'account_number_confirmation',
+  'routing_number',
+  'account_type',
+  'name_on_account',
+  'country'
 ];
 
 /**
@@ -78,7 +77,7 @@ var requiredFields = [
  */
 
 function token (options, done) {
-  var data = normalize(options);
+  var data = normalize(fields, options);
   var input = data.values;
   var userErrors = validate(input);
 
@@ -127,32 +126,6 @@ function bankInfo (options, done) {
     if (err) return done(err);
     done(null, res);
   });
-}
-
-/**
- * Parses options out of a form element and normalizes according to rules.
- *
- * @param {Object|HTMLFormElement} options
- * @return {Object}
- */
-
-function normalize (options) {
-  var el = dom.element(options);
-  var data = { fields: {}, values: {} };
-
-  if (el && 'form' === el.nodeName.toLowerCase()) {
-    each(el.querySelectorAll('[data-recurly]'), function (field) {
-      var name = dom.data(field, 'recurly');
-      if (~index(fields, name)) {
-        data.fields[name] = field;
-        data.values[name] = dom.value(field);
-      }
-    });
-  } else {
-    data.values = options;
-  }
-
-  return data;
 }
 
 /**
