+++ lib/recurly/pricing/index.js
@@ -4,9 +4,7 @@
 
 var Emitter = require('emitter');
 var index = require('indexof');
-var each = require('each');
 var type = require('type');
-var bind = require('bind');
 var find = require('find');
 var mixin = require('mixin');
 var keys = require('object').keys;
@@ -43,11 +41,11 @@ Emitter(Pricing.prototype);
  */
 
 Pricing.properties = [
-    'plan'
-  , 'addon'
-  , 'coupon'
-  , 'address'
-  , 'currency'
+  'plan',
+  'addon',
+  'coupon',
+  'address',
+  'currency'
 ];
 
 /**
@@ -92,13 +90,13 @@ Pricing.prototype.remove = function (opts, done) {
         item = self.items.addons.splice(pos);
       }
     } else if (self.items[prop] && (id === self.items[prop].code || id === true)) {
-      item = self.items[prop]
+      item = self.items[prop];
       delete self.items[prop];
     } else {
       return reject(errors('unremovable-item', {
-          type: prop
-        , id: id
-        , reason: 'does not exist on this pricing instance.'
+        type: prop,
+        id: id,
+        reason: 'does not exist on this pricing instance.'
       }));
     }
   }, this).nodeify(done);
@@ -337,15 +335,15 @@ Pricing.prototype.tax = function (tax, done) {
 
 Pricing.prototype.currency = function (code, done) {
   var self = this;
-  var plan = this.items.plan
+  var plan = this.items.plan;
   var currency = this.items.currency;
 
   return new PricingPromise(function (resolve, reject) {
     if (currency === code) return resolve(currency);
     if (plan && !(code in plan.price)) {
       return reject(errors('invalid-currency', {
-          currencyCode: code
-        , planCurrencies: keys(plan.price)
+        currencyCode: code,
+        planCurrencies: keys(plan.price)
       }));
     }
