+++ lib/recurly/token.js
@@ -3,13 +3,11 @@
  * Module dependencies.
  */
 
-var bind = require('bind');
 var each = require('each');
 var type = require('type');
 var index = require('indexof');
 var debug = require('debug')('recurly:token');
-var dom = require('../util/dom');
-var parseCard = require('../util/parse-card');
+var normalize = require('../util/normalize');
 var errors = require('../errors');
 
 /**
@@ -26,21 +24,21 @@ module.exports = token;
  */
 
 var fields = [
-    'first_name'
-  , 'last_name'
-  , 'number'
-  , 'month'
-  , 'year'
-  , 'cvv'
-  , 'address1'
-  , 'address2'
-  , 'country'
-  , 'city'
-  , 'state'
-  , 'postal_code'
-  , 'phone'
-  , 'vat_number'
-  , 'token'
+  'first_name',
+  'last_name',
+  'number',
+  'month',
+  'year',
+  'cvv',
+  'address1',
+  'address2',
+  'country',
+  'city',
+  'state',
+  'postal_code',
+  'phone',
+  'vat_number',
+  'token'
 ];
 
 /**
@@ -69,8 +67,7 @@ var fields = [
  */
 
 function token (options, done) {
-  var open = bind(this, this.open);
-  var data = normalize(options);
+  var data = normalize(fields, options, { parseCard: true });
   var input = data.values;
   var userErrors = validate.call(this, input);
 
@@ -89,34 +86,6 @@ function token (options, done) {
     }
     done(null, res);
   });
-};
-
-/**
- * Parses options out of a form element and normalizes according to rules.
- *
- * @param {Object|HTMLFormElement} options
- * @return {Object}
- */
-
-function normalize (options) {
-  var el = dom.element(options);
-  var data = { fields: {}, values: {} };
-
-  if (el && 'form' === el.nodeName.toLowerCase()) {
-    each(el.querySelectorAll('[data-recurly]'), function (field) {
-      var name = dom.data(field, 'recurly');
-      if (~index(fields, name)) {
-        data.fields[name] = field;
-        data.values[name] = dom.value(field);
-      }
-    });
-  } else {
-    data.values = options;
-  }
-
-  data.values.number = parseCard(data.values.number);
-
-  return data;
 }
 
 /**
